Exclude career update messages from chat notifications
This commit modifies the `BroadcastCareerUpdateJob` and `Message` models to exclude career update messages from chat notifications. The `BroadcastCareerUpdateJob` class now retrieves the messages for the sender using the `messages` association, and then excludes any messages that have a `type` attribute of `'CareerUpdate'`. This ensures that career update messages are not included in the list of messages that are sent to supporters and subscribers.

The `Message` model now includes a `type` attribute that allows messages to be filtered by type. Career update messages are given a type of `'CareerUpdate'`, which allows them to be excluded from chat notifications.

These changes improve the user experience by ensuring that career update messages do not clutter chat notifications, while still allowing supporters and subscribers to receive important updates from the sender.

TO-DO:
[  ] Testing